mqtt-network-monitor-addon/requirements.txt
fastapi>=0.104.0,<1.0
uvicorn>=0.24.0,<1.0
paho-mqtt>=2.0,<3.0
